feat: move StoryColoredProfileAvatar to /components/avatar
