+++ lib/app/components/avatar/story_colored_profile_avatar.dart
